requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==2.2.5
django-celery-beat==2.0.0
django-timezone-field==4.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
